authentication logic working
