+++ backend/src/main/java/com/lernia/auth/model/User.java
@@ -1,9 +1,22 @@
 package com.lernia.auth.model;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+
+@Entity
 public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     private String username;
     private String password;
+    private String email; 
+
 
     public User() {
     }
@@ -12,7 +25,9 @@ public class User {
         this.id = id;
         this.username = username;
         this.password = password;
+        this.email = email;
     }
+
 
     public Long getId() {
         return id;
@@ -36,5 +51,13 @@ public class User {
 
     public void setPassword(String password) {
         this.password = password;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
     }
 }
