+++ backend/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/LerniaDB
+spring.datasource.url=jdbc:postgresql://localhost:5432/lerniadb
 spring.datasource.username=user1
 spring.datasource.password=pass1
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
