+++ backend/src/test/java/com/lernia/auth/AuthServiceTests.java
@@ -1,0 +1,36 @@
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+
+import com.lernia.auth.repository.UserRepository;
+
+class AuthServiceTests {
+
+    @InjectMocks
+    private AuthService authService;
+
+    @Mock
+    private UserRepository userRepository;
+
+    @BeforeEach
+    void setUp() {
+        MockitoAnnotations.openMocks(this);
+    }
+
+    @Test
+    void testLoginSuccess() {
+        // This test will be implemented later
+        //fail("Not yet implemented");
+        assertTrue(true);
+    }
+
+    @Test
+    void testLoginFailure() {
+        // This test will be implemented later
+        assertTrue(true);
+        //fail("Not yet implemented");
+    }
+}
