--- backend/src/test/java/com/lernia/login/AuthServiceTests.java
@@ -1,32 +1,0 @@
-import static org.junit.jupiter.api.Assertions.*;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-
-class AuthServiceTests {
-
-    @InjectMocks
-    private AuthService authService;
-
-    @Mock
-    private UserRepository userRepository;
-
-    @BeforeEach
-    void setUp() {
-        MockitoAnnotations.openMocks(this);
-    }
-
-    @Test
-    void testLoginSuccess() {
-        // This test will be implemented later
-        fail("Not yet implemented");
-    }
-
-    @Test
-    void testLoginFailure() {
-        // This test will be implemented later
-        fail("Not yet implemented");
-    }
-}
